+++ vs code extension/client/src/services/cutLanguageClientServerSetup.ts
@@ -1,0 +1,67 @@
+import * as path from 'path';
+import { workspace, ExtensionContext, window} from 'vscode';
+
+import {
+	LanguageClient,
+	LanguageClientOptions,
+	ServerOptions,
+	TransportKind,
+} from 'vscode-languageclient/node';
+
+import * as LOGGER from '../utils/Logger'
+
+let client: LanguageClient;
+
+export function startCutLanguageClientServer(context: ExtensionContext) {
+	// The server is implemented in node
+	let serverModule = context.asAbsolutePath(
+		path.join('server', 'out', 'server.js')
+	);
+	// The debug options for the server
+	// --inspect=6009: runs the server in Node's Inspector mode so VS Code can attach to the server for debugging
+	let debugOptions = { execArgv: ['--nolazy', '--inspect=6009'] };
+
+	// If the extension is launched in debug mode then the debug server options are used
+	// Otherwise the run options are used
+	let serverOptions: ServerOptions = {
+		run: { module: serverModule, transport: TransportKind.ipc },
+		debug: {
+			module: serverModule,
+			transport: TransportKind.ipc,
+			options: debugOptions
+		}
+	};
+
+	// Options to control the language client
+	let clientOptions: LanguageClientOptions = {
+		// Register the server for cut documents
+		documentSelector: [{ scheme: 'file', language: 'cut' }],
+		synchronize: {
+			// Notify the server about file changes to '.clientrc files contained in the workspace
+			fileEvents: workspace.createFileSystemWatcher('**/.clientrc')
+		},
+		outputChannel: window.createOutputChannel('Cut Language Extension'),
+	};
+
+	// Create the language client and start the client.
+	client = new LanguageClient(
+		'cut',
+		'CUT',
+		serverOptions,
+		clientOptions
+	);
+
+	//Log information to a dedicated outputchannel
+	LOGGER.createLogger(clientOptions.outputChannel)
+	LOGGER.log('Starting Client', LOGGER.INFO)
+	
+	// Start the client. This will also launch the server
+	client.start();
+}
+
+export function stopCutLanguageClientServer(): Thenable<void> | undefined {
+	if (!client) {
+		return undefined;
+	}
+	return client.stop();
+}