+++ vs code extension/CHANGELOG.md
@@ -11,3 +11,11 @@ All notable changes to the "cobol-unit-test" extension will be documented in thi
 - Updated Readme, as not all information was correct
 - Added icon
 - Cleaned up some unnecessary files
+
+## [0.2.0] 04.03.2022
+
+- Integrated Cobol Check into extension
+- Added Cobol Check view
+- Added command for running Cobol Check
+- Added command for configuring Cobol Check
+- Added command for resetting the configurations